Definition Updates
@DavidCrosswellGSQ @GSQ-AI @ajtroup @geoderekh @LukeHauck 
All, I have updated the definitions section in this branch to better delineate between the classes in the model. These are imperfect and open to debate but I hope it kicks off a way to re-align with SOSA and create clear boundaries and definitions. I have also included a practical usage section for implementation and what can and can't be inferred instead of explicitly detailing. Please read the definitions section to discuss on Monday.